requirements.txt
pymongo
python-dotenv
pandas
os
